refactor(test): split change proof tests into focused files
Split the monolithic mod.rs (2073 lines) into 9 focused test files,
each with a module doc comment describing its purpose:

  structural.rs      — one test per error variant in verify_change_proof_structure
  root_hash.rs       — root hash verification happy paths (one per proof config)
  attack.rs          — forged, omitted, or spurious batch ops
  edge_cases.rs      — odd depths, divergence, exclusion, deletion, children
  truncation.rs      — truncation, generator, boundary ops, disjoint proofs
  divergent_child.rs — prove() divergent child inclusion in exclusion proofs
  regression.rs      — regression tests for specific bug fixes
  empty_and_bounded.rs — empty start trie and bounded range verification
  pipeline.rs        — architectural code path tests (Steps 2-4, ConflictingProofNodes,
                       ValueAtOddNibbleLength, EndProofOperationMismatch, MissingEndProof)

mod.rs now contains only shared infrastructure (~90 lines): new_db,
setup_db!, setup_2nd_commit!, verify_and_check, and mod declarations.

No test logic changed. 67 non-slow tests pass, zero clippy warnings.
